[DP-516] 기술블로그 검색어 조회시 score 누락 이슈 수정

+++ src/main/java/com/dreamypatisiel/devdevdev/web/controller/techArticle/TechArticleController.java
@@ -37,7 +37,7 @@ public class TechArticleController {
             @RequestParam(required = false) Long techArticleId,
             @RequestParam(required = false) String keyword,
             @RequestParam(required = false) Long companyId,
-            @RequestParam(required = false) Float score
+            @RequestParam(required = false) Double score
     ) {
         TechArticleService techArticleService = techArticleServiceStrategy.getTechArticleService();
         Authentication authentication = AuthenticationMemberUtils.getAuthentication();
